pyproject.toml
[project]
name = "biblion-management-system"
version = "0.0.1"
description = "A very Simple Book Management System"
authors = [
    {name = "Bittu Kumar",email = "bitweb24@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12.3"
dependencies = [
    "annotated-types (==0.7.0)",
    "anyio (==4.8.0)",
    "certifi (==2025.1.31)",
    "click (==8.1.8)",
    "dnspython (==2.7.0)",
    "ecdsa (==0.19.0)",
    "email-validator (==2.2.0)",
    "fastapi (==0.115.8)",
    "fastapi-cli (==0.0.7)",
    "greenlet (==3.1.1)",
    "h11 (==0.14.0)",
    "httpcore (==1.0.7)",
    "httptools (==0.6.4)",
    "httpx (==0.28.1)",
    "idna (==3.10)",
    "jinja2 (==3.1.5)",
    "markdown-it-py (==3.0.0)",
    "markupsafe (==3.0.2)",
    "mdurl (==0.1.2)",
    "passlib (==1.7.4)",
    "pyasn1 (==0.4.8)",
    "pydantic (==2.10.6)",
    "pydantic-core (==2.27.2)",
    "pygments (==2.19.1)",
    "python-dotenv (==1.0.1)",
    "python-jose (==3.4.0)",
    "python-multipart (==0.0.20)",
    "pyyaml (==6.0.2)",
    "rich (==13.9.4)",
    "rich-toolkit (==0.13.2)",
    "rsa (==4.9)",
    "shellingham (==1.5.4)",
    "six (==1.17.0)",
    "sniffio (==1.3.1)",
    "sqlalchemy (==2.0.38)",
    "starlette (==0.45.3)",
    "typer (==0.15.1)",
    "typing-extensions (==4.12.2)",
    "uvicorn (==0.34.0)",
    "uvloop (==0.21.0)",
    "watchfiles (==1.0.4)",
    "websockets (==15.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pylint = "^3.3.4"
ruff = "^0.9.9"

[tool.ruff]
line-length = 88
indent-width = 4
target-version = "py312"

[tool.ruff.lint]
ignore = ["E501", "D"]
select = ["E4", "E7", "E9", "F"]

# Allow fix for all enabled rules (when `--fix`) is provided.
fixable = ["ALL"]
unfixable = []

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

# Enable auto-formatting of code examples in docstrings. Markdown,
# reStructuredText code/literal blocks and doctests are all supported.
#
# This is currently disabled by default, but it is planned for this
# to be opt-out in the future.
docstring-code-format = false

# Set the line length limit used when formatting code snippets in
# docstrings.
#
# This only has an effect when the `docstring-code-format` setting is
# enabled.
docstring-code-line-length = "dynamic"


[tool.pylint.master]
rcfile = ".pylintrc"

